Select/Where/ForEach expanded to use -Object suffix. This eliminates Invoke-ScriptAnalyzer warnings. An override was added to TabExpansion to avoid a tonne of warnings there which were irrelevant.
Empty/dummy catch blocks were expanded to provide a warning.

Get-Alias was changed to use a Where-Object in order to avoid some
pollution of Error[].

SetIsReceived had a duplicate and both functions were misnamed. These
were merged.

Changes were made to improve the possibility of using Set-StrictMode.
* Where arrays are checked for .Count the array gets checked directly
  instead.
* Where variables aren't defined they are defined.
* Where properties are used they are checked first, especially on
  objects which are expected to be an array but sometimes aren't.

PSEdition doesn't exist on PS4 and so it has an extra check before
testing the property.

TabExpansion fails on its function override on extremely high load
(the property disappears). So it has been wrapped in a wildcard test.

Start-RSJob used an incorrect Finished setting. This was fixed to
match a first usage of it.

An If block was merged in Get-RSJob as it wasn't needed.

A 100ms wait was removed from Wait-RSJob as it can check for a
Completed property instead. This fixes a race condition on jobs
being passed onto Receive-RSJob before they're ready with data.