Validation of custom module(s) path to import.
When the '-ModulesFromPathToImport' parameter is passed for
'Start-RSJob', it should handle either a full and relative path name.

Also raise an exception if the module(s) path is not found, which means
none of the module(s) are imported in the runspace job. An exception
prevents the runspace jobs from starting, as any dependent functions
referenced inside the 'ScriptBlock' parameter would not be found.