Consolidate module import into single parameter.
Remove separate '-ModulesFromPathToImport' parameter, adding this
functionality to existing '-ModulesToImport' parameters. This means
'-ModulesToImport' can accept a collection containing:

 - Installed module names
 - Path (full or relative) containing one or more modules
 - Path (full or relative) to a module manifest, script module, or
   binary module file

Purpose of consolidation is to simplify module import for the caller;
They should have the flexibility to pass a root module path for
recursive import, or be able to specify modules similar to `Import-Module`,
with minimal preprocessing of the parameter collection.